pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "direct-transcriber"
version = "0.1.0"
description = "Local CPU-only batch audio transcription tool for RAGflow"
authors = [{name = "User", email = "user@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "openai-whisper>=20231117",
    "click>=8.0",
    "rich>=13.0",
    "ffmpeg-python>=0.2.0",
    "psutil>=5.9.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy>=1.21.0",
]

[project.scripts]
direct-transcriber = "direct_transcriber.cli:main"

[project.urls]
Homepage = "https://github.com/user/direct-transcriber"

[tool.hatch.build.targets.wheel]
packages = ["src/direct_transcriber"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/pyproject.toml",
]